# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Escolar",
        "description": "API para sistema de gerenciamento escolar multi-tenant. Permite cadastro de usuários, alunos, turmas, diários, eventos, entre outros. O contexto do tenant é definido pelo header 'x-tenant-id'.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "alexseyf66@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Servidor Local"
        },
        {
            "url": "https://elo-api-alexseyfs-projects.vercel.app",
            "description": "Servidor de Produção"
        }
    ],
...
    "paths": {
...
        "/alunos": {
            "get": {
                "tags": ["Alunos"],
                "summary": "Listar alunos",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Lista de alunos" }
                }
            },
            "post": {
                "tags": ["Alunos"],
                "summary": "Criar aluno",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": {
                        "application/json": { "schema": { "$ref": "#/components/schemas/AlunoInput" } }
                    }
                },
                "responses": {
                    "201": { "description": "Aluno criado" },
                    "400": { "description": "Dados inválidos" }
                }
            }
        },
        "/alunos/ativos": {
            "get": {
                "tags": ["Alunos"],
                "summary": "Listar alunos ativos",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Lista de alunos ativos" }
                }
            }
        },
        "/alunos/{id}": {
            "get": {
                "tags": ["Alunos"],
                "summary": "Obter aluno por ID",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Dados do aluno" },
                    "404": { "description": "Aluno não encontrado" }
                }
            },
            "put": {
                "tags": ["Alunos"],
                "summary": "Atualizar aluno (Completo)",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AlunoInput" } } }
                },
                "responses": { "200": { "description": "Aluno atualizado" } }
            },
            "patch": {
                "tags": ["Alunos"],
                "summary": "Atualizar aluno (Parcial)",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AlunoInput" } } }
                },
                "responses": { "200": { "description": "Aluno atualizado" } }
            }
        },
        "/alunos/soft/{id}": {
            "delete": {
                "tags": ["Alunos"],
                "summary": "Desativar aluno (Soft Delete)",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Aluno desativado" } }
            }
        },
        "/alunos/{usuarioId}/responsavel": {
            "post": {
                "tags": ["Alunos"],
                "summary": "Adicionar responsável ao aluno",
                "parameters": [
                    { "name": "usuarioId", "in": "path", "description": "ID do usuário responsável", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": { "alunoId": { "type": "integer" } },
                                "required": ["alunoId"]
                            }
                        }
                    }
                },
                "responses": { "201": { "description": "Vínculo criado" } }
            }
        },
        "/alunos/{alunoId}/responsavel/{usuarioId}": {
            "delete": {
                "tags": ["Alunos"],
                "summary": "Remover responsável do aluno",
                "parameters": [
                    { "name": "alunoId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "name": "usuarioId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Responsável removido" } }
            }
        },
        "/alunos/relatorios/mensalidades-por-turma": {
            "get": {
                "tags": ["Alunos"],
                "summary": "Relatório de mensalidades por turma",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Relatório gerado" } }
            }
        },
        "/alunos/{alunoId}/possui-registro-diario": {
            "get": {
                "tags": ["Alunos"],
                "summary": "Verificar se aluno possui diário na data",
                "parameters": [
                    { "name": "alunoId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "name": "data", "in": "query", "schema": { "type": "string", "format": "date" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Status verificado" } }
            }
        },
...
        "/turmas/{turmaId}/alunos": {
            "get": {
                "tags": ["Turmas"],
                "summary": "Listar alunos da turma",
                "parameters": [
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de alunos" } }
            }
        },
...
        "/responsaveis/{responsavelId}/alunos": {
            "get": {
                "tags": ["Responsaveis"],
                "summary": "Alunos de um responsável",
                "parameters": [
                    { "name": "responsavelId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de alunos" } }
            }
        },
...
        "/diarios/aluno/{alunoId}": {
            "get": {
                "tags": ["Diarios"],
                "summary": "Diários de um aluno",
                "parameters": [
                    { "name": "alunoId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de diários" } }
            }
        },
...
    },
    "components": {
        "parameters": {
            "TenantId": {
                "name": "x-tenant-id",
                "in": "header",
                "description": "Slug ou ID da escola (Tenant)",
                "required": true,
                "schema": { "type": "string" }
            }
        },
        "schemas": {
...
            "AlunoInput": {
                "type": "object",
                "properties": {
                    "nome": { "type": "string" },
                    "dataNasc": { "type": "string", "format": "date-time" },
                    "turmaId": { "type": "integer" },
                    "isAtivo": { "type": "boolean" }
                },
                "required": ["nome", "dataNasc"]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}